New FillInForm test focusing on the stack overflow question
Insert double Value for equotion in PDF with pdfbox in Java
https://stackoverflow.com/questions/64912688/insert-double-value-for-equotion-in-pdf-with-pdfbox-in-java
2020-04_Beschaffungsantrag.pdf
https://1drv.ms/b/s!Av6exjPNXlgOioouAuXL6QV4eUGkqg?e=ocfhvC

testFill2020_04BeschaffungsantragEinzelpreis - The field number 30 used by the OP is not the terminal field to fill-in but merely an intermediary one. The field to actually use is 'Einzelpreis in € exkl USt1.0.0.0.0'. Furthermore, as the field list used by the OP, it does not contain this field. Instead the field should be retrieved by name from the PDAcroForm object.