New RefreshAppearances tests focusing on the stack overflow question 65665511
Added font not loading correct pdfbox for acroform
https://stackoverflow.com/questions/65701597/added-font-not-loading-correct-pdfbox-for-acroform
DejaVuSans.ttf
https://www.fontsquirrel.com/fonts/DejaVu-Sans
sample5.pdf
https://pdfhost.io/v/Oh9~qVoG5_sample5pdf.pdf

testRefreshLikeLuckydonald - This test essentially calls the OP's code. Debugging made clear that the current AcroForm form field refreshing mechanism in PDFBox is not really usable in combination with fonts yet to be subsetted. The underlying problem can be reduced to what is shown in the test testIllustrateType0Issue().

testIllustrateType0Issue - This test illustrates the problem one can run into when using a PDFBox PDType0Font with subsetting activated: If one does not always use the font directly but instead looks it up in the resources, a new PDType0Font is generated from the preliminary PDF objects created from the original instance, but the contents thereof make this new instance appear non-embedded, so no information are given to the original instance which glyphs are to be embedded and need to be in the ToUnicode mapping. This is exactly the problem the PDFBox code for refreshing AcroForm appearances runs into in testRefreshLikeLuckydonald().