New OverlayDocuments tests focusing on the stack overflow question 66122899
Is it possible to “repair” a pdf using pdfbox?
https://stackoverflow.com/questions/66122899/is-it-possible-to-repair-a-pdf-using-pdfbox
example_broken.pdf
https://drive.google.com/file/d/10lYNfkQlUvxeZ2rFps2ozBO-zgGrnVeU/view?usp=sharing

testOverlayPreparationExampleBroken - After the OP shared his code, it turns out that not the overlaying (tested in testOverlayExampleBroken) but instead a preparation step is the problem: the page resources in the document are inherited but the preparation step loses everything inherited.

testOverlayPreparationFixedExampleBroken - The problem identified in testOverlayPreparationExampleBroken can be fixed by explicitly setting the page resources as shown here.