New MergeXObjectIntoPage test focusing on the stack overflow question 66274818
Add FormXobject content from resources to content stream using PDFBox?
https://stackoverflow.com/questions/66274818/add-formxobject-content-from-resources-to-content-stream-using-pdfbox
HighPioneerFallNewsletterADApdf_2.pdf
https://drive.google.com/file/d/1lT4pwGAk4I6vaekTHhWnx_hngcKCm0DK/view?usp=sharing

testMergeLikeFascinatingCoder - This test takes the OP's code and adds a q-Q envelope around the inserted instructions. This doesn't make a difference for the example document because in it the XObject Do instruction always was the last instruction before a Q anyways. The problem with the example document is based in the XObjects being transparency groups and the copy of the instructions does not transport the grouping.