New InnerImport test focusing on the stack overflow question 68213927
PDFBox : small pdf file can't be open in PDFormXObject
https://stackoverflow.com/questions/68213927/pdfbox-small-pdf-file-cant-be-open-in-pdformxobject
Pour Ronald - PDF Autocad reduit.pdf
https://wetransfer.com/downloads/583ee85f7d433f2b523fe27dea6df48520210701163459/b21644

testImportAsFormLikeGuyard - Indeed, importPageAsForm first calls importOcProperties which clones the source OCGs into the target OCGs using PDFCloneUtility. As source and target are the same, something the utility does not expect, this results in an endless loop adding the contents of an array to itself. A check for equality at the top of cloneMerge can prevent this.