New EditMarkedContent test focusing on the stack overflow question 68565744
PDFBox: how to determine bounding box of vector figure (path shape)
https://stackoverflow.com/questions/68565744/pdfbox-how-to-determine-bounding-box-of-vector-figure-path-shape
test.pdf as "testAlexanderDyuzhev.pdf"
https://www.dropbox.com/s/w4ksnud78bu9oz5/test.pdf?dl=0

MarkedContentBoundingBoxFinder - This PDFGraphicsStreamEngine subclass allows determining the bounding boxes of marked content according to their instructions inside. It is derived from the BoundingBoxFinder and, therefore, also is subject to the restrictions and limitations of that class.

testMarkedContentTestAlexanderDyuzhev - This test applies the MarkedContentBoundingBoxFinder to the test document provided by the OP and frames the bounding boxes accordingly.

testMarkedContentTrade_Simple1 - This test applies the MarkedContentBoundingBoxFinder to a test document of some other question and frames the bounding boxes accordingly.

BoundingBoxFinder - getCurrentPoint returns a non-null value to prevent incorrect logging messages by the line-to listener.