New ScalePageWithAnnots test focusing on the stack overflow question 72349310
How to scale a PDAnnotation?
https://stackoverflow.com/questions/72349310/how-to-scale-a-pdannotation
sample.pdf as "sampleSignedByAlias.pdf"
https://github.com/amberream/Resources/raw/master/sample.pdf

testForPereZix - This test extends the OP's code to also scale the rectangle of the annotations. Beware: For a complete solution more has to be scaled. On one hand there are annotation types with additional coordinate fields like QuadPoints and on the other hand there are additional properties to scale in particular for annotations without appearance streams, like border width, font size, ...