New RemoteSigning test focusing on the stack overflow question 75505900
pdfbox - document getting corrupted after adding signed attributes
https://stackoverflow.com/questions/75505900/pdfbox-document-getting-corrupted-after-adding-signed-attributes

testSignLikeSkdjksDfksl - This is essentially the OP's code after he added a message digest attribute. It still doesn't work because the signature still signs the document hash, not the signed attributes.

testSignLikeSkdjksDfkslImproved - This is an improved version of the OP's code in testSignLikeSkdjksDfksl. Now the signature signs the signed attributes, not only the document hash.