New AddAnnotationIncrementally test focusing on the stack overflow question 76027267
Signing a pdf digitally multiple times and adding text/images to it while signing
https://stackoverflow.com/questions/76027267/signing-a-pdf-digitally-multiple-times-and-adding-text-images-to-it-while-signin
sample-signed-first.pdf (first revision thereof as sample.pdf)
https://drive.google.com/file/d/11h9aQLcjGpJKdJTwLNEl3ax4hbkxI6iw/view?usp=share_link
PDFBox - opening and saving a signed pdf invalidates my signature
https://stackoverflow.com/questions/41467415/pdfbox-opening-and-saving-a-signed-pdf-invalidates-my-signature

testAnnotateRawatrSample - This test uses the code in EDIT 3 of the latter question to add a stamp annotation to the example PDF as incremental update like the author of the former question claims to have done. But the error he observed cannot be reproduced.