New IsolateFigures tests focusing on the stack overflow question 76333586
images inverted and split when extracting images from pdf document by using PDFBox or Poppler
https://stackoverflow.com/questions/76333586/images-inverted-and-split-when-extracting-images-from-pdf-document-by-using-pdfb
my.pdf
https://banyafx.oss-cn-hangzhou.aliyuncs.com/assets/pdf/my.pdf

testIsolateInMy:

Based on the initial observation that the figures in the documents are drawn without text object insets, this code isolates the figures by writing each block between text objects to a separate page.

Obviously, there are some such blocks that only consist of e.g. an underline. The code attempts to drop them, see the conditional addition of the new page to the document in endFigurePage.

As it turns out, there unfortunately are some figures which appear to have been patched afterwards; those patches then are displayed on separate pages...