New AddLink test focusing on the stack overflow question 76738878
Add PDAnnotationLink to a PDF With pdfbox
https://stackoverflow.com/questions/76738878/add-pdannotationlink-to-a-pdf-with-pdfbox

testAddLinkLikeMarioRovelli - Indeed, the text under the link annotation is half covered by a black rectangle. Looking closely at the source code one can see what this black rectangle is: The code requests an underline of width 10! Considering that the font size of the text underneath also is 10, it is not surprising to see half covered...