New RemoveText tests focusing on the stack overflow question 77992635
How to remove specific pattern text from PDF using PDFBox?
https://stackoverflow.com/questions/77992635/how-to-remove-specific-pattern-text-from-pdf-using-pdfbox
sample_tag.pdf
https://github.com/nofelkad/pdf-sample/blob/main/sample_tag.pdf
text_tag.pdf
https://github.com/nofelkad/pdf-sample/blob/main/text_tag.pdf

removeTagsLikeVasK - This essentially is the code provided by the OP with only COSString decoding and encoding fixed.

testRemoveTagsFromSampleTag - As the OP already said, his code posted as "Update 1" works as expected for this file, even after the corrected decoding and encoding.

testRemoveTagsFromTextTag - As the OP already said, his code posted as "Update 1" works does not work for this file, instead it scrambled the displayed text. After correcting the decoding and encoding of COS string data, this problem stopped. It became apparent, though, that the OP's routines to filter the COS string contents are deficient.